requirements.txt
streamlit==1.44.1
google-cloud-bigquery==3.31.0
db-dtypes==1.4.2